+++ src/data/records/mod.rs
@@ -2,7 +2,10 @@ use std::{error::Error, fs::File, io::ErrorKind, path::Path};
 
 use crate::{
     data::fasta::{FastaAA, FastaNT, FastaNTAnnot, FastaSeq},
-    prelude::{FastQ, FastQView, FastQViewMut},
+    prelude::{
+        AminoAcids, AminoAcidsView, AminoAcidsViewMut, FastQ, FastQView, FastQViewMut, Nucleotides, NucleotidesView,
+        NucleotidesViewMut,
+    },
 };
 
 /// A module for reading and manipulating
@@ -162,5 +165,106 @@ impl HeaderReadable for FastaNTAnnot {
     #[inline]
     fn header(&self) -> &str {
         &self.name
+    }
+}
+
+/// Getter trait for structures providing read access to a sequences.
+///
+/// The sequence can be either nucleotides or amino acids, and is returned as a
+/// byte slice.
+pub trait SequenceReadable {
+    /// Get the sequence from the struct as a byte slice.
+    #[must_use]
+    fn sequence_bytes(&self) -> &[u8];
+}
+
+impl SequenceReadable for Nucleotides {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for NucleotidesView<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for NucleotidesViewMut<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for AminoAcids {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for AminoAcidsView<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for AminoAcidsViewMut<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.as_ref()
+    }
+}
+
+impl SequenceReadable for FastQ {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastQView<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastQViewMut<'_> {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastaSeq {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastaAA {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastaNT {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
+    }
+}
+
+impl SequenceReadable for FastaNTAnnot {
+    #[inline]
+    fn sequence_bytes(&self) -> &[u8] {
+        self.sequence.as_ref()
     }
 }