+++ CHANGELOG.md
@@ -14,6 +14,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added `to_eq_x` to `Alignment` and `AlignmentStates`, which converts `M` to either `=` or `X` in the alignment
 - Added `from_ciglets_unchecked` and `from_cigar_unchecked` to `AlignmentStates`
 - Added `decode_iter` to `KmerEncoder`
+- Added `SequenceReadable` getter trait to record types and sequence types
 - Added `nw_score_from_path` for rescoring a global alignment (behind `alignment-diagnostics` feature gate)
 - Introduced the `AlignmentAndSeqs` arbitrary wrapper (behind `fuzzing` feature gate)
 - Added `visit_params` methods to pHMMs as a diagnostic tool for inspecting the parameters that are visited along a given path (behind `dev-phmm` and `alignment-diagnostics` feature gates)